Call get points API , closed #8

+++ app/src/main/java/com/example/mapboxexample/data/model/PointServer.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class PointServer(
     @SerializedName("id") val id: Long?,
-    @SerializedName("latitude") val latitude: Float?,
-    @SerializedName("longitude") val longitude: Float?
+    @SerializedName("latitude") val latitude: Double?,
+    @SerializedName("longitude") val longitude: Double?
 )
